refactored the apis to use dependency io
